+++ middleware/session-timeout.js
@@ -4,7 +4,7 @@ module.exports = (app, proxyUrl, waypoints, sessionTtl, timeoutDialogCountdown) 
 
     res.locals.timeoutDialog = {
       keepAliveUrl: `${mountUrl}${waypoints.SESSION_KEEP_ALIVE}`,
-      signOutUrl: `${mountUrl}${waypoints.SESSION_RESTART}`,
+      signOutUrl: `${mountUrl}?ref=sign-out`,
       timeoutUrl: `${mountUrl}${waypoints.SESSION_ENDED}`,
       countdown: sessionTtl - timeoutDialogCountdown,
       timeout: sessionTtl