+++ routes/confirmation.js
@@ -4,14 +4,11 @@ const appVersion = require('../package.json').version;
 module.exports = function (casaApp) {
   const { router } = casaApp
   router.get('/confirmation', function (req, res) {
-    const { transactionID } = req.session.downloadContext
-
     //  Clear the journey data once submitted
     casaApp.endSession(req).then(() => {
       res.render('confirmation', {
         sessionId: encodeURIComponent(req.session.id),
-        appVersion: appVersion,
-        transactionID
+        appVersion: appVersion
       });
     }).catch((err) => {
       logger.error('Session ending error - ' + err);